add admin users

+++ database/migrations/2020_03_25_163446_add_fields_in_admins_table.php
@@ -16,7 +16,7 @@ class AddFieldsInAdminsTable extends Migration
         Schema::table('admins', function (Blueprint $table) {
             $table->renameColumn('name', 'first_name');
             $table->string('last_name')->after('name')->nullable();
-            $table->integer('mobile_number')->after('email')->nullable();
+            $table->bigInteger('mobile_number')->after('email')->nullable();
         });
     }
